modify opd codes..

+++ src/main/java/com/hims/entity/repository/ProcedureDetailsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProcedureDetailsRepository extends JpaRepository<ProcedureDetails, Integer> {
+public interface ProcedureDetailsRepository extends JpaRepository<ProcedureDetails, Long> {
 }
